Rename controller classes
InformativeController was renamed into ObjectiveValueDisplayController
and EmptyController was renamed into ProgressBarController.
