+++ src/main/resources/com/targus/objectiveValueDisplay.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.targus.ui.controllers.InformativeController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.targus.ui.controllers.ObjectiveValueDisplayController">
     <VBox prefHeight="141.0" prefWidth="378.0" spacing="15.0">
         <HBox alignment="CENTER_LEFT" prefHeight="30.0" prefWidth="380.0" spacing="15.0">
             <Label prefHeight="18.0" prefWidth="126.0" text="Sensor objective" />
